[BugFix] Use old sort_key_idxes in to construct seekTuple after light schema change (#36830)
Why I'm doing:
If a pk table execute a light schema change, causing the column index of the sort key to change,
it may cause a crash when querying the table.
This problem occurs for two reasons:
1. When querying the table, it use the sort key column offset is to create seektuples. The new schema
metadata send by FE is not used here. It means that we use old version of tablet_schema.
2. However, even if the latest schema information send by FE is used, it will still cause a crash.
This is because the variable that saves the latest schema information from fe does not update the sort
key idxes information.

What I'm doing:
1. use new tablet_schema to construct seektuples.
2. reset sort key idxes information when construct the new tablet_schema from FE

Signed-off-by: srlch <linzichao@starrocks.com>