[BugFix] Fix miss update job state when replaying restore log may cause restored table lost after FE restart
Why I'm doing:
If we set history_job_keep_max_second < checkpoint time interval (no matter by background daemon
or manlly create image). When we replay the log for the restore job, the final FINISHED state log
will be skipped to update the state job state in BackupHandler because this job is considered
as a expired job controlled by history_job_keep_max_second.

But the point here is that, Only FINISHED state is skipped to update but not other state log, the BackupHandler
will persistent the unfinished state for this job even the job is actually finished now.

When we restart FE in this case, we will get a unfinished restore job in FE and want to rewrite the original
table which is restored before FE restart and cause this problem.

What I'm doing:
When we meet a expired job with FINISHED state, we remove the info info existed in BackupHandler.

Signed-off-by: srlch <linzichao@starrocks.com>